Move Type assertions to Var
