+++ src/Assertion/Boolean/IsFalse.php
@@ -14,7 +14,7 @@ use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
 use Vivarium\Assertion\String\IsEmpty;
-use Vivarium\Assertion\Type\IsBoolean;
+use Vivarium\Assertion\Var\IsBoolean;
 
 use function sprintf;
