+++ src/Assertion/Conditional/Each.php
@@ -14,7 +14,7 @@ use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
 use Vivarium\Assertion\String\IsEmpty;
-use Vivarium\Assertion\Type\IsArray;
+use Vivarium\Assertion\Var\IsArray;
 
 use function sprintf;
