+++ src/Assertion/Encoding/IsSystemEncoding.php
@@ -14,8 +14,8 @@ use ValueError;
 use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
-use Vivarium\Assertion\Type\IsBoolean;
-use Vivarium\Assertion\Type\IsString;
+use Vivarium\Assertion\Var\IsBoolean;
+use Vivarium\Assertion\Var\IsString;
 
 use function mb_internal_encoding;
 use function sprintf;
