+++ src/Assertion/String/IsLongAtLeast.php
@@ -15,7 +15,7 @@ use Vivarium\Assertion\Encoding\IsSystemEncoding;
 use Vivarium\Assertion\Exception\AssertionFailed;
 use Vivarium\Assertion\Helpers\TypeToString;
 use Vivarium\Assertion\Numeric\IsGreaterThan;
-use Vivarium\Assertion\Type\IsString;
+use Vivarium\Assertion\Var\IsString;
 
 use function mb_strlen;
 use function sprintf;
