+++ src/Assertion/Var/IsCallable.php
@@ -8,7 +8,7 @@
 
 declare(strict_types=1);
 
-namespace Vivarium\Assertion\Type;
+namespace Vivarium\Assertion\Var;
 
 use Vivarium\Assertion\Assertion;
 use Vivarium\Assertion\Exception\AssertionFailed;
@@ -16,19 +16,19 @@ use Vivarium\Assertion\Helpers\TypeToString;
 use Vivarium\Assertion\String\IsEmpty;
 
 use function gettype;
-use function is_string;
+use function is_callable;
 use function sprintf;
 
-/** @template-implements Assertion<string> */
-final class IsString implements Assertion
+/** @template-implements Assertion<callable> */
+final class IsCallable implements Assertion
 {
-    /** @psalm-assert string $value */
+    /** @psalm-assert callable $value */
     public function assert(mixed $value, string $message = ''): void
     {
         if (! $this($value)) {
             $message = sprintf(
                 ! (new IsEmpty())($message) ?
-                     $message : 'Expected value to be string. Got %2$s.',
+                     $message : 'Expected value to be callable. Got %2$s.',
                 (new TypeToString())($value),
                 gettype($value),
             );
@@ -37,9 +37,9 @@ final class IsString implements Assertion
         }
     }
 
-    /** @psalm-assert-if-true string $value */
+    /** @psalm-assert-if-true callable $value */
     public function __invoke(mixed $value): bool
     {
-        return is_string($value);
+        return is_callable($value);
     }
 }
